Actividades hecho

+++ src/main/java/services/ActivityService.java
@@ -1,7 +1,9 @@
 
 package services;
 
+import java.util.Calendar;
 import java.util.Collection;
+import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,6 +13,7 @@ import org.springframework.util.Assert;
 import repositories.ActivityRepository;
 import domain.Activity;
 import domain.Actor;
+import domain.Conference;
 
 @Service
 @Transactional
@@ -23,6 +26,9 @@ public class ActivityService {
 	// Supporting services
 	@Autowired
 	private ActorService		actorService;
+
+	@Autowired
+	private ConferenceService	conferenceService;
 
 
 	// Simple CRUD methods
@@ -46,27 +52,47 @@ public class ActivityService {
 		return result;
 	}
 
-	public Activity save(final Activity activity) {
+	//R14.6
+	public Activity save(final Activity activity, final Conference conference) {
 		Assert.notNull(activity);
-
-		final Actor actorLogged = this.actorService.findActorLogged();
-		Assert.notNull(actorLogged);
 
 		Activity result;
 
-		if (activity.getId() == 0)
+		final Actor actorLogged = this.actorService.findActorLogged();
+		Assert.notNull(actorLogged);
+		this.actorService.checkUserLoginAdministrator(actorLogged);
+
+		final Calendar endDate = Calendar.getInstance();
+		endDate.setTime(conference.getEndDate());
+		endDate.add(Calendar.DATE, 1);
+
+		Assert.isTrue(conference.getIsFinalMode(), "Activities can only be managed if the conference is in final mode");
+		Assert.isTrue(conference.getStartDate().compareTo(new Date(System.currentTimeMillis())) > 0, "Activities can only be managed if the start date of the conference has not passed");
+		Assert.isTrue(activity.getStartMoment().compareTo(conference.getStartDate()) >= 0 && activity.getStartMoment().compareTo(endDate.getTime()) < 0,
+			"The start moment of an activity must be equal to or later than the start date of the conference and lower than the end date of the conference");
+
+		if (activity.getId() == 0) {
 			result = this.activityRepository.save(activity);
-		else
+			conference.getActivities().add(result);
+			this.conferenceService.saveAuxiliar(conference);
+		} else
 			result = this.activityRepository.save(activity);
 
 		return result;
 	}
 
-	public void delete(final Activity activity) {
+	//R14.6
+	public void delete(final Activity activity, final Conference conference) {
 		Assert.notNull(activity);
 		Assert.isTrue(activity.getId() != 0);
 		Assert.isTrue(this.activityRepository.exists(activity.getId()));
 
+		final Actor actorLogged = this.actorService.findActorLogged();
+		Assert.notNull(actorLogged);
+		this.actorService.checkUserLoginAdministrator(actorLogged);
+
+		conference.getActivities().remove(activity);
+		this.conferenceService.saveAuxiliar(conference);
 		this.activityRepository.delete(activity);
 	}
