+++ src/main/java/services/SectionService.java
@@ -11,6 +11,7 @@ import org.springframework.util.Assert;
 import repositories.SectionRepository;
 import domain.Actor;
 import domain.Section;
+import domain.Tutorial;
 
 @Service
 @Transactional
@@ -24,13 +25,18 @@ public class SectionService {
 	@Autowired
 	private ActorService		actorService;
 
+	@Autowired
+	private TutorialService		tutorialService;
+
 
 	// Simple CRUD methods
+	//R14.6
 	public Section create() {
 		Section result;
 
 		final Actor actorLogged = this.actorService.findActorLogged();
 		Assert.notNull(actorLogged);
+		this.actorService.checkUserLoginAdministrator(actorLogged);
 
 		result = new Section();
 
@@ -57,30 +63,56 @@ public class SectionService {
 		return result;
 	}
 
-	public Section save(final Section section) {
+	//R14.6
+	public Section save(final Section section, final Tutorial tutorial) {
 		Assert.notNull(section);
 
 		final Actor actorLogged = this.actorService.findActorLogged();
 		Assert.notNull(actorLogged);
+		this.actorService.checkUserLoginAdministrator(actorLogged);
 
 		Section result;
 
-		if (section.getId() == 0)
+		if (section.getId() == 0) {
 			result = this.sectionRepository.save(section);
-		else
+			tutorial.getSections().add(result);
+			this.tutorialService.saveAuxiliar(tutorial);
+		} else
 			result = this.sectionRepository.save(section);
 
 		return result;
 	}
 
-	public void delete(final Section section) {
+	//R14.6
+	public void delete(final Section section, final Tutorial tutorial) {
 		Assert.notNull(section);
 		Assert.isTrue(section.getId() != 0);
 		Assert.isTrue(this.sectionRepository.exists(section.getId()));
+
+		final Actor actorLogged = this.actorService.findActorLogged();
+		Assert.notNull(actorLogged);
+		this.actorService.checkUserLoginAdministrator(actorLogged);
+
+		final Collection<Section> sectionsTutorial = tutorial.getSections();
+		Assert.isTrue(sectionsTutorial.size() != 1, "The tutorial must be at least one section");
+
+		sectionsTutorial.remove(section);
+		tutorial.setSections(sectionsTutorial);
+		this.tutorialService.saveAuxiliar(tutorial);
 
 		this.sectionRepository.delete(section);
 	}
 
 	// Other business methods
+	public Collection<Section> findSectionsByTutorial(final int tutorialId) {
+		Collection<Section> result;
 
+		final Actor actorLogged = this.actorService.findActorLogged();
+		Assert.notNull(actorLogged);
+		this.actorService.checkUserLoginAdministrator(actorLogged);
+
+		result = this.sectionRepository.findSectionsByTutorialId(tutorialId);
+
+		return result;
+	}
 }
